CFix potentential arithmetic overflows detected by msvc static analysis
git-svn-id: svn://scribus.net/branches/Version15x/Scribus@25385 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/ui/hruler.h
@@ -76,7 +76,7 @@ public:
 	double offset() const { return m_offset; }
 	
 private:
-	int findRulerHandle(QPoint mp, double grabRadius);
+	int findRulerHandle(QPoint mp, int grabRadius);
 	
 	virtual void paintEvent(QPaintEvent *e);
 	virtual void mousePressEvent(QMouseEvent *m);
